Add files via upload
TicTacToc is an Android app that allows two people to play the game on a phone, with the additional feature that each play has to be done within a limited amount of time. if a player does not play within the time limit they lose their turn. The app keeps track of the players' scores - the first player to win 5 games is the winner of the tournament. The following details are easy to understand if you download, install, and run the TicTacToc app.

To run the application download the two "MainActivity" files, which define the main code of the application, then download the two "activity_main" files which defines the layout of the app. 

Note: Adding all the files of the application was not feasible since it amounts to 700+ files including the gradle build folders. Therefore, I added these four files, which detail the main functionality of the code, and showcases my logic behind coding a functional tic-tac-toe application.